first proof of concept
README:
this is hacked together in a really short time and thus quite buggy.
however it already shows how the "dynamic toolbar/menu" stuff will work.

try opening muse, creating some part in the arranger, opening a ScoreEdit
and after that a PianoRoll (in that order!). Then arrange the windows so
you can see and access both, and switch between them around. You will see
that parts of the menu bar, and also the toolbars will change.

I'm planning to make the toolbars remember their state (hidden or not,
position) on a per-window-class basis, and to somehow make the menus more
similar (that is, the menu bar itself won't change when changing windows,
only the menus change). The arranger will be a dedicated window (then it
will regain its controls again ;) ).